OneDrive/Área de Trabalho/microservicos/boilerplate/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0  # for tests

# Auth
python-jose[cryptography]==3.3.0
httpx==0.27.0  # Google OAuth exchange

# Billing
stripe==9.9.0

# Email
resend==2.1.0

# Utilities
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
